Fix java packages
