+++ com.ibm.streamsx.iot/impl/java/src/com/ibm/streamsx/iot/DeviceCmd.java
@@ -2,73 +2,81 @@
 # Licensed Materials - Property of IBM
 # Copyright IBM Corp. 2016 
  */
-package com.ibm.streamsx.iotf;
+package com.ibm.streamsx.iot;
 
 import java.io.IOException;
 import java.io.Serializable;
 
 import com.ibm.json.java.JSON;
 import com.ibm.json.java.JSONObject;
+import com.ibm.streams.operator.OutputTuple;
 import com.ibm.streams.operator.Tuple;
 import com.ibm.streamsx.topology.tuple.JSONAble;
 
 /**
- * A device event.
+ * A device command.
  */
-public class DeviceEvent implements JSONAble, Serializable {
-    public static final String EVENT_ID = "eventId";
-    
+public class DeviceCmd implements JSONAble, Serializable {
     private static final long serialVersionUID = 1L;
     
+    public static final String CMD_ID = "cmdId";
+    
     private final Device device;
-    private final String eventId;
-    private final String ts;
+    private final String cmdId;
     private final JSONObject payload;
     
-    public DeviceEvent(Device device, String eventId, JSONObject data) {
+    public DeviceCmd(Device device, String cmdId, JSONObject payload) {
         super();
         this.device = device;
-        this.eventId = eventId;
-        this.ts = (String) data.get("ts");
-        this.payload = (JSONObject) data.get("d");
-        
+        this.cmdId = cmdId;
+        this.payload = payload;      
     }
-    public static DeviceEvent newDeviceEvent(Tuple tuple) {
-        return newDeviceEvent(
+    
+    public static DeviceCmd newDeviceCmd(Tuple tuple) {
+        return newDeviceCmd(
                 Device.newDevice(tuple),
-                tuple.getString(EVENT_ID),
+                tuple.getString(CMD_ID),
                 tuple.getString(Device.JSON));
     }
-    public static DeviceEvent newDeviceEvent(Device device, String eventId, String serializedData) {      
+    public static DeviceCmd newDeviceCmd(Device device, String cmdId, String serializedData) {      
         JSONObject payload;
         try {
             payload = (JSONObject) JSON.parse(serializedData);
         } catch (IOException e) {
             throw new RuntimeException(e);
         }        
-        return new DeviceEvent(device, eventId, payload);       
+        return new DeviceCmd(device, cmdId, payload);       
+    }
+   
+    
+    public OutputTuple toTuple(OutputTuple tuple) {
+        tuple.setString(Device.TYPE_ID, getDevice().getTypeId());
+        tuple.setString(Device.DEVICE_ID, getDevice().getId());
+        tuple.setString(DeviceCmd.CMD_ID, getCmdId());
+        try {
+            tuple.setString(Device.JSON, getPayload().serialize());
+        } catch (Exception e) {
+            throw new RuntimeException(e);
+        }
+        return tuple;
     }
 
     public Device getDevice() {
         return device;
     }
 
-    public String getEventId() {
-        return eventId;
+    public String getCmdId() {
+        return cmdId;
     }
 
     public JSONObject getPayload() {
         return payload;
     }
-    public String getTs() {
-        return ts;
-    }
     
     @Override
     public JSONObject toJSON() {
         JSONObject json = getDevice().toJSON();
-        json.put(EVENT_ID, getEventId());
-        json.put("ts", getTs());
+        json.put("cmdId", getCmdId());
         json.put("payload", getPayload());
         return json;
     }
@@ -80,7 +88,4 @@ public class DeviceEvent implements JSONAble, Serializable {
             throw new RuntimeException(e);
         }
     }
-    
-    
-    
 }
