+++ com.ibm.streamsx.iot/impl/java/src/com/ibm/streamsx/iot/package-info.java
@@ -5,4 +5,4 @@
 /**
  * Java application API support for connecting to IBM Watson IoT Platform. 
  */
-package com.ibm.streamsx.iotf;
+package com.ibm.streamsx.iot;
